Moved backend testing logic to its own file
